fixed gcc compile error

+++ libtorrent/src/hash-sha1.c
@@ -1,6 +1,6 @@
 #include "sha.h"
 
-int hash_sha1(const uint8_t* data, size_t bytes, uint8_t sha1[20])
+int hash_sha1(const uint8_t* data, unsigned int bytes, uint8_t sha1[20])
 {
 	SHA1Context ctx;
 	SHA1Reset(&ctx);
